Fixed seeds along the analysis and added extra map variance analysis
Now the seeds of the stochastic part of the analysis (simulation production) can be fixed. Also, they are shared between different lmax runs, thus one can apprecciate better the influence of lmax on the results. Furhermore, I have added the analysis to compute the extra variance on the sky-map of the reconstructed full-sky alms.